Moved various setup code to new files Fixes #36
