+++ functions/coord_grid.php
@@ -1,0 +1,10 @@
+<?php
+
+class Coord_Grid
+{
+	const SIZE_LAT = 'SIZE_LAT';
+	const SIZE_LON = 'SIZE_LON';
+	public array $params = array();
+}
+
+?>
