+++ functions/make_fasta_header.php
@@ -1,5 +1,7 @@
 <?php
 
+require_once $FUNCTIONS_DIR. 'bold.php';
+
 function make_fasta_header(array $seq_data) {
 
     if (($id = $seq_data[BOLD::PROCESS_ID]) == '') {
