+++ functions/total_sequence_count.php
@@ -1,6 +1,7 @@
 <?php 
 
-include_once $FUNCTIONS_DIR. 'sequence_sets.php';
+require_once $FUNCTIONS_DIR. 'sequence_sets.php';
+require_once $FUNCTIONS_DIR. 'taxsets.php';
 
 // lookup total_sequence_count field in the data file
 function total_sequence_count($taxon) {
